requirements.txt
# Core dependencies
pyautogen>=0.2.0
openai>=1.0.0

# Web framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Configuration
python-dotenv>=1.0.0

# Utilities
typing-extensions>=4.8.0
